fixed an issue with the pvc name

+++ kubernetes/apps/volsync-system/volsync-src-fixer/app/resources/volsync-src-fixer.sh
@@ -1,8 +1,8 @@
 #!/bin/bash
 set -e
 
-function LOG() {
-    echo "[$(date +'%Y-%m-%d %H:%M:%S')] $1"
+LOG() {
+    echo "[$(date +'%Y-%m-%d %H:%M:%S')] ${*}"
 }
 
 get_pods() {
@@ -10,15 +10,23 @@ get_pods() {
       jq -c '.items[] | select(.metadata.name | startswith("volsync-src")) | [.metadata.name, .metadata.namespace, .metadata.labels."job-name"]' | \
       sed -e 's/\[//g' -e 's/\]//g' -e 's/\"//g'
 }
+
+LOG "🔍Checking for stuck volsync-src pods in all namespaces"
+
 PODS=($(get_pods))
-for POD in "${PODS[@]}"; do
-    POD=(${POD//,/ })
-    POD_NAME="${POD[0]}"
-    NAMESPACE="${POD[1]}"
-    PVC_NAME="${POD[2]}"
-    echo "Found stuck pod '$POD_NAME' in namespace '$NAMESPACE' using pvc '$PVC_NAME'"
-    kubectl -n "$NAMESPACE" delete pvc "$PVC_NAME" --wait=false
-    LOG "Deleted pvc '$PVC_NAME'"
-    kubectl -n "$NAMESPACE" delete pod "$POD_NAME"
-    LOG "Deleted pod '$POD_NAME in namespace '$NAMESPACE'"
-done
+if [ ${#PODS[@]} -eq 0 ]; then
+    LOG "🦄No volsync-src pods found"
+else
+    for POD in "${PODS[@]}"; do
+        POD=(${POD//,/ })
+        POD_NAME="${POD[0]}"
+        NAMESPACE="${POD[1]}"
+        JOB_NAME="${POD[2]}"
+        PVC_NAME="${JOB_NAME//-src/}-src"
+        LOG "💢Found stuck pod '$POD_NAME' in namespace '$NAMESPACE' using pvc '$PVC_NAME'"
+        kubectl -n "$NAMESPACE" delete pvc "$PVC_NAME" --wait=false
+        LOG "💥Deleted pvc '$PVC_NAME' in namespace '$NAMESPACE'"
+        kubectl -n "$NAMESPACE" delete pod "$POD_NAME"
+        LOG "💥Deleted pod '$POD_NAME in namespace '$NAMESPACE'"
+    done
+fi
